changed sql server login info to accomodate invoke-sqlcmd.  Added a function to update daily values.

+++ 10a Eaton Creek/10a_telemetry.ps1
@@ -19,7 +19,9 @@ $Server = $config."ftp-url"
 $ServerDir = $config."ftp-dir"
 
 # SQL
-$SQLConnString = $config."sql-connStr"
+$SQLServer = $config."sql-server"
+$SQLDB = $config."sql-database"
+$SQLConnString = "Data Source=$SQLServer; Database=$SQLDB; Trusted_Connection=True;"
 
 # Local
 $LocalDir = $config."temp-dir"
@@ -37,6 +39,9 @@ $webclient.Credentials = New-Object System.Net.NetworkCredential($Username, $Pas
 ######################
 # Load data from files
 ######################
+
+$MinDate = @()
+$MaxDate = @()
 
 foreach ($file in $FilesToDownload) {
     $RemoteFile = "$Server$ServerDir/$file"
@@ -57,13 +62,17 @@ foreach ($file in $FilesToDownload) {
     $mapTemp = New-InSituMap -param 'temperature'
     $mapPressure = New-InSituMap -param 'pressure'
 
+    # First, record the minimum and maximum dates in the file
+    $MinDate += $table[1].Compute("Min(DateTime)", "Provisional=1")
+    $MaxDate += $table[1].Compute("Max(DateTime)", "Provisional=1")
+
     Write-Verbose "Uploading data to GData..."
-    Import-TableToSQL -InTable $table[1] -ConnString $SQLConnString -DestTable "tblDischargeGauging" -mapping $mapStage
-    Import-TableToSQL -InTable $table[1] -ConnString $SQLConnString -DestTable "tblWaterTempGauging" -mapping $mapTemp
-    Import-TableToSQL -InTable $table[1] -ConnString $SQLConnString -DestTable "tblBarometerGauging" -mapping $mapPressure
+    #Import-TableToSQL -InTable $table[1] -ConnString $SQLConnString -DestTable "tblDischargeGauging" -mapping $mapStage
+    #Import-TableToSQL -InTable $table[1] -ConnString $SQLConnString -DestTable "tblWaterTempGauging" -mapping $mapTemp
+    #Import-TableToSQL -InTable $table[1] -ConnString $SQLConnString -DestTable "tblBarometerGauging" -mapping $mapPressure
 
     Write-Verbose "Deleting $RemoteFile..."
-    Remove-FTPFile -User $Username -Pass $Password -Svr $Server -FilePath "$ServerDir/$file"
+    #Remove-FTPFile -User $Username -Pass $Password -Svr $Server -FilePath "$ServerDir/$file"
 
     Write-Verbose "Adding file to archive..."
     & "$(Split-Path $PSScriptRoot)\7z\7za.exe" a archive.zip $LocalFile
@@ -73,6 +82,34 @@ foreach ($file in $FilesToDownload) {
 # Cleanup
 #########
 
-Write-Verbose "Deleting local files from $LocalDir"
-Remove-Item "$($LocalDir)*.*"
 
+if ($MinDate.Length -eq 0) {
+    Write-Verbose "No files found on FTP site; nothing to do."
+} else {
+
+    $FirstDateTime = ($MinDate | Sort-Object)[0]
+    $LastDateTime  = ($MaxDate | Sort-Object)[-1]
+
+    $MinDate | Out-String | Write-Debug
+    $MaxDate | Out-String | Write-Debug
+
+    $queryParams = @{
+        'SQLServer'  = $SQLServer;
+        'SQLDB'      = $SQLDB;
+        'table'      = 'tblDischargeDaily';
+        'gid'        = '965';
+        'BeginDate'  = $FirstDateTime;
+        'EndDate'    = $LastDateTime
+    }
+
+    $tables = 'tblDischargeDaily', 'tblTempDaily', 'tblBarometerDaily'
+
+    Write-Verbose "Updating daily tables from $FirstDateTime to $LastDateTime..."
+    foreach ($t in $tables) {
+        $queryParams['table'] = $t
+        Update-DailySQLTable @queryParams
+    }
+
+    Write-Verbose "Deleting local files from $LocalDir"
+    Remove-Item "$($LocalDir)*.*"
+}
